Move unit test and nightMode url blocking function
