+++ ios/brave-ios/Sources/Brave/Frontend/Browser/Tab.swift
@@ -414,7 +414,9 @@ class Tab: NSObject {
     didSet {
       var isNightModeEnabled = false
 
-      if let fetchedTabURL = fetchedURL, nightMode, !fetchedTabURL.isNightModeBlockedURL {
+      if let fetchedTabURL = fetchedURL, nightMode,
+        !DarkReaderScriptHandler.isNightModeBlockedURL(fetchedTabURL)
+      {
         isNightModeEnabled = true
       }
