+++ ios/brave-ios/Sources/Brave/Frontend/UserContent/UserScripts/Scripts_Dynamic/ScriptHandlers/Sandboxed/DarkReaderScriptHandler.swift
@@ -100,9 +100,48 @@ public class DarkReaderScriptHandler: TabContentScript {
     Preferences.General.nightModeEnabled.value = enabled
 
     for tab in tabManager.allTabs {
-      if let fetchedTabURL = tab.fetchedURL, !fetchedTabURL.isNightModeBlockedURL {
+      if let fetchedTabURL = tab.fetchedURL, !isNightModeBlockedURL(fetchedTabURL) {
         tab.nightMode = enabled
       }
     }
+  }
+}
+
+extension DarkReaderScriptHandler {
+  // Check if the website is a night mode blocked site
+  public static func isNightModeBlockedURL(_ url: URL) -> Bool {
+    // The reason we use `normalizedHost` is because we want to keep the eTLD+1
+    // IE: (search.brave.com instead of brave.com)
+    guard let urlHost = url.normalizedHost(), let registry = url.publicSuffix else {
+      return false
+    }
+
+    // Remove the `registry` so we get `search.brave` instead of `search.brave.com`
+    // We get amazon instead of amazon.co.uk
+    // mail.proton instead of mail.proton.com
+    let domainName = urlHost.dropLast(registry.count).trimmingCharacters(
+      in: CharacterSet(charactersIn: ".")
+    )
+
+    // Site domains that should NOT inject night mode
+    let siteList = Set([
+      "twitter", "youtube", "twitch",
+      "soundcloud", "github", "netflix",
+      "imdb", "mail.proton", "amazon",
+      "x", "search.brave",
+
+      // Search Engines
+      "search.brave", "google", "qwant",
+      "startpage", "duckduckgo", "presearch",
+
+      // Dev sites
+      "macrumors", "9to5mac", "developer.apple",
+
+      // Casual sites
+      "wowhead", "xbox", "thegamer",
+      "cineplex", "starwars",
+    ])
+
+    return siteList.contains(domainName)
   }
 }
