+++ ios/brave-ios/Tests/BraveSharedTests/URLExtensionTests.swift
@@ -207,64 +207,6 @@ class URLExtensionTests: XCTestCase {
     XCTAssertFalse(ntpTestURL.isShredAvailable)
   }
 
-  func testNightModeBlockedURL() {
-    let blockList = [
-      "twitter.com",
-      "m.twitter.com",
-      "youtube.com",
-      "m.youtube.com",
-      "www.youtube.com",
-      "amazon.co.uk",
-      "amazon.com",
-      "m.amazon.com",
-      "m.amazon.co.uk",
-      "x.com",
-      "m.x.com",
-      "search.brave.com",
-      "m.search.brave.com",
-      "www.search.brave.com",
-    ]
-
-    let allowList = [
-      "twitter.brave.com",
-      "m.twitter.brave.com",
-      "m.search.amazon.com",
-      "m.search.amazon.co.uk",
-      "search.amazon.co.uk",
-      "search.amazon.brave.com",
-      "m.brave.com",
-      "brave.com",
-      "talk.brave.com",
-      "bbc.co.uk",
-      "news.bbc.co.uk",
-      "bbc.example.co.uk",
-      "example.bbc.foo.co.uk",
-      "example.bbc.foo.com",
-    ]
-
-    for urlString in blockList {
-      guard let url = URL(string: "https://\(urlString)") else {
-        XCTFail("Invalid URL: \(urlString)")
-        continue
-      }
-
-      if !url.isNightModeBlockedURL {
-        print("URL FAILED NIGHTMODE: \(urlString)")
-      }
-
-      XCTAssertTrue(url.isNightModeBlockedURL)
-    }
-
-    for urlString in allowList {
-      guard let url = URL(string: "https://\(urlString)") else {
-        XCTFail("Invalid URL: \(urlString)")
-        continue
-      }
-
-      XCTAssertFalse(url.isNightModeBlockedURL)
-    }
-  }
-
   func testBlobURLAuthStripping() {
     var url = URL(string: "blob:https://brave.github.io/1945c9fb-9834-479b-bb3a-e67e3a9408cf")!
     XCTAssertEqual(
